Create a documentation link for blocks

Creates a styled external link that opens block documentation in a new tab. Used to provide easy access to reference documentation and guides from within blocks.

Usage

doc_link(
  text,
  url,
  icon_name = "arrow-up-right-from-square",
  tooltip = NULL,
  class = ""
)

Arguments

text

Character string for the link text

url

Character string for the documentation URL

icon_name

Character string for Font Awesome icon name (default: "arrow-up-right-from-square")

tooltip

Optional character string for hover tooltip

class

Additional CSS classes to add

Value

An HTML link element with icon

Examples

if (FALSE) { # \dontrun{
# Basic usage
doc_link(
  "Documentation",
  "https://bristolmyerssquibb.github.io/blockr.dplyr/reference/new_mutate_block.html"
)

# With tooltip
doc_link(
  "Expression helpers",
  "https://bristolmyerssquibb.github.io/blockr.dplyr/articles/expression-helpers.html",
  tooltip = "Learn about common functions"
)
} # }